io_antarctica_scene: More fixes to get kart and track exporter working; not 100% there
More issues have been found with both exporters, and they've been fixed.
Additional changes include:
* io_scene_spm: Fix MixRGB shader node creation
* Kart exporter now deselects all objects before performing any mesh exporting
